moved task counter to its own class
