+++ MoodSpace/app/src/main/java/com/example/moodspace/FollowController.java
@@ -378,42 +378,3 @@ public class FollowController implements ControllerCallback {
         cc.callback(callbackId, bundle);
     }
 }
-
-class CompletedTaskCounter {
-    private int limit;
-    private int complete = 0;
-    private int success = 0;
-
-    public CompletedTaskCounter(int limit) {
-        if (limit <= 0) {
-            throw new IllegalArgumentException("inputted limit should be greater than 0");
-        }
-        this.limit = limit;
-    }
-
-    public static CompletedTaskCounter getDefault() {
-        return new CompletedTaskCounter(2);
-    }
-
-    public void incrementComplete() {
-        if (complete + 1 > limit) {
-            throw new IllegalArgumentException("Did not expect to increment complete past max");
-        }
-        complete++;
-    }
-
-    public void incrementSuccess() {
-        if (success + 1 > limit) {
-            throw new IllegalArgumentException("Did not expect to increment success past max");
-        }
-        success++;
-    }
-
-    public boolean isComplete() {
-        return complete == limit;
-    }
-
-    public boolean isSuccess() {
-        return success == limit;
-    }
-}
